backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
attrs==25.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
faiss-cpu==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.4
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langsmith==0.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
orjson==3.10.15
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
PyJWT==2.8.0
pypdf==5.3.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.4.4
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
yarl==1.18.3
zstandard==0.23.0
